requirements.txt
streamlit==1.32.0
torchaudio==2.2.2
scikit-learn=1.2.2
numpy==1.23.5
pytorch==2.2.2
pillow==10.4.0
